Fix bug where empty sections list doesn't show coteacher invitation
