+++ apps/src/templates/studioHomepages/teacherHomepageV2/SectionList.tsx
@@ -23,12 +23,10 @@ import _ from 'lodash';
 import React, {useState} from 'react';
 
 import Spinner from '@cdo/apps/sharedComponents/Spinner';
-import {AgeGatedSectionsBanner} from '@cdo/apps/templates/policy_compliance/AgeGatedSectionsModal/AgeGatedSectionsBanner';
 import {
   removeSectionOrThrow,
   setSectionOrder,
 } from '@cdo/apps/templates/teacherDashboard/teacherSectionsRedux';
-import {atRiskAgeGatedSections} from '@cdo/apps/templates/teacherDashboard/teacherSectionsReduxSelectors';
 import {SectionMap} from '@cdo/apps/templates/teacherDashboard/types/teacherSectionTypes';
 import HttpClient from '@cdo/apps/util/HttpClient';
 import {useAppSelector, useAppDispatch} from '@cdo/apps/util/reduxHooks';
@@ -37,7 +35,6 @@ import {
   getFilteredSectionOrderIds,
   saveSectionOrder,
 } from '../../teacherDashboard/sectionOrderUtils';
-import CoteacherInviteNotification from '../CoteacherInviteNotification';
 
 import {SectionCard} from './SectionCard';
 import {SectionDeleteModal} from './SectionDeleteModal';
@@ -68,17 +65,6 @@ export const SectionList: React.FC<SectionListProps> = ({
   showHiddenOnly,
 }) => {
   const dispatch = useAppDispatch();
-
-  const [CAPmodalOpen, setCAPModalOpen] = React.useState(false);
-  const toggleCAPModal = () => {
-    setCAPModalOpen(!CAPmodalOpen);
-  };
-
-  const ageGatedSections = useAppSelector(atRiskAgeGatedSections);
-
-  const shouldDisplayAtRiskAgeGatedWarning = () => {
-    return ageGatedSections?.length > 0;
-  };
 
   const [sectionToDelete, setSectionToDelete] = useState<number>(NO_SECTION_ID);
   const sections: SectionMap = useAppSelector(
@@ -177,41 +163,31 @@ export const SectionList: React.FC<SectionListProps> = ({
   return (
     <div id="ui-test-section-list">
       {sectionsAreLoaded ? (
-        <>
-          {shouldDisplayAtRiskAgeGatedWarning() && (
-            <AgeGatedSectionsBanner
-              toggleModal={toggleCAPModal}
-              modalOpen={CAPmodalOpen}
-              ageGatedSections={ageGatedSections}
-            />
-          )}
-          <CoteacherInviteNotification isForPl={false} destructiveLoad={true} />
-          <DndContext
-            sensors={sensors}
-            collisionDetection={closestCenter}
-            onDragEnd={handleDragEnd}
-            modifiers={[restrictToVerticalAxis, restrictToParentElement]}
+        <DndContext
+          sensors={sensors}
+          collisionDetection={closestCenter}
+          onDragEnd={handleDragEnd}
+          modifiers={[restrictToVerticalAxis, restrictToParentElement]}
+        >
+          <SortableContext
+            items={sortableSectionIds}
+            strategy={verticalListSortingStrategy}
           >
-            <SortableContext
-              items={sortableSectionIds}
-              strategy={verticalListSortingStrategy}
-            >
-              <ol className={styles.sectionList}>
-                {sectionIdsToShow.map(id =>
-                  sections[id] ? (
-                    <SectionCard
-                      id={id}
-                      key={id}
-                      section={sections[id]}
-                      onDeleteClickCallback={onDeleteClickCallback}
-                      studioUrlPrefix={studioUrlPrefix}
-                    />
-                  ) : null
-                )}
-              </ol>
-            </SortableContext>
-          </DndContext>
-        </>
+            <ol className={styles.sectionList}>
+              {sectionIdsToShow.map(id =>
+                sections[id] ? (
+                  <SectionCard
+                    id={id}
+                    key={id}
+                    section={sections[id]}
+                    onDeleteClickCallback={onDeleteClickCallback}
+                    studioUrlPrefix={studioUrlPrefix}
+                  />
+                ) : null
+              )}
+            </ol>
+          </SortableContext>
+        </DndContext>
       ) : (
         <Spinner size="large" />
       )}
